✨ feat(user): migrer vers ApiPlatform v4 — remplacer UserDataPersister par UserProcessor
- Remplace le persister legacy `App\DataPersister\UserDataPersister` (namespace `ApiPlatform\Core\*`) par
  `App\State\Processor\UserProcessor` (implémente `ApiPlatform\State\ProcessorInterface`).
- Fait migrer la logique de hash du mot de passe et de persist/flush dans le Processor (fichier ajouté :
  `src/State/Processor/UserProcessor.php`).
- Ajoute un shim déprécié `src/DataPersister/UserDataPersister.php` qui lève une exception explicite pour
  éviter tout usage accidentel tant que la migration n'est pas validée.
- Motivation : compatibilité avec API Platform v4 (nouvelle architecture state/processor), résolution des
  erreurs d'analyse statique/IDE et usage correct de l'autoconfiguration Symfony.
- Actions post-merge / vérifications :
  - Lancer `composer install` si nécessaire, puis `php ./bin/phpunit --testdox`.
  - Tester manuellement : POST `/api/users` → vérifier que le mot de passe est hashé en base.
  - Supprimer le shim legacy une fois la migration validée et nettoyer la configuration si besoin.
- Risques / notes :
  - Peut nécessiter un ajustement des services si vous aviez des références explicites à l'ancien persister.
  - Aucun changement de contrat API attendu, juste déplacement de la logique serveur.
